+++ cacheman/lib/cacheman.dart
@@ -1,8 +1,7 @@
 export 'src/cacheman.dart';
 export 'src/debug.dart';
-export 'src/engine.dart' show CacheOptions, CachemanOptions, Engine;
 export 'src/entity.dart' show CacheEntity, defaultSerialize, defaultDeserialize;
 export 'src/fast.dart';
-export 'src/interface.dart' show Store, MemoCache, Codec, CachemanOnError;
+export 'src/interface.dart' show Codec, CachemanOnError;
 export 'src/jsonx.dart' show Jsonx;
-export 'src/memo.dart' show Memo;
+export 'src/options.dart' show CacheOptions, CachemanOptions;